Modificando seeders para creacion de conceptos necesarios par admision

+++ database/migrations/2021_06_29_120926_create_detalles_admision_table.php
@@ -17,7 +17,7 @@ class CreateDetallesAdmisionTable extends Migration
             $table->id();
             $table->tinyInteger('concepto_id');
             $table->bigInteger('admision_id');
-            $table->bigInteger('precio_variable')->nullable();
+            $table->double('precio_variable')->nullable();
             $table->timestamps();
             $table->softDeletes();
